requirements.txt
flask==2.3.6
flask-cors==3.1.2
torch==2.1.0
torchvision==0.16.1
pillow==10.0.1
